# gh-102578: Optimise setting and deleting mutable attributes on non-dataclass subclasses of frozen dataclasses