--- a/Lib/dataclasses.py
+++ b/Lib/dataclasses.py
@@ -1,4 +1,3 @@
-from asyncio import Condition
 import re
 import sys
 import copy