--- a/Lib/test/test_dataclasses.py
+++ b/Lib/test/test_dataclasses.py
@@ -2779,8 +2779,6 @@ class TestFrozen(unittest.TestCase):
         self.assertFalse(hasattr(c, 'i'))
         with self.assertRaises(FrozenInstanceError):
             del c.i
-        with self.assertRaises(AttributeError):
-            del c.i
 
     def test_inherit(self):
         @dataclass(frozen=True)
@@ -2904,6 +2902,35 @@ class TestFrozen(unittest.TestCase):
         self.assertEqual(s.x, 3)
         self.assertEqual(s.y, 10)
         self.assertEqual(s.cached, True)
+
+        with self.assertRaises(FrozenInstanceError):
+            del s.x
+        self.assertEqual(s.x, 3)
+        with self.assertRaises(FrozenInstanceError):
+            del s.y
+        self.assertEqual(s.y, 10)
+        del s.cached
+        self.assertFalse(hasattr(s, 'cached'))
+        with self.assertRaises(AttributeError):
+            del s.cached
+
+    def test_non_frozen_normal_derived_from_empty_frozen(self):
+        @dataclass(frozen=True)
+        class D:
+            pass
+
+        class S(D):
+            pass
+
+        s = S()
+        self.assertFalse(hasattr(s, 'x'))
+        s.x = 5
+        self.assertEqual(s.x, 5)
+
+        del s.x
+        self.assertFalse(hasattr(s, 'x'))
+        with self.assertRaises(AttributeError):
+            del s.x
 
     def test_overwriting_frozen(self):
         # frozen uses __setattr__ and __delattr__.