--- a/Misc/NEWS.d/next/Library/2023-03-10-10-10-00.gh-issue-102573.GQ7sV6.rst
+++ b/Misc/NEWS.d/next/Library/2023-03-10-10-10-00.gh-issue-102573.GQ7sV6.rst
@@ -1,1 +1,0 @@
-Use ``set``-based name lookup rather than ``tuple``\s for frozen dataclasses