--- a/Misc/NEWS.d/next/Library/2023-03-10-13-21-16.gh-issue-102578.-gujoI.rst
+++ b/Misc/NEWS.d/next/Library/2023-03-10-13-21-16.gh-issue-102578.-gujoI.rst
@@ -1,0 +1,2 @@
+Speed up sanity check for ``__setattr__`` and ``__delattr__`` for frozen
+``dataclass``\es by using ``set``-based name lookup rather than ``tuple``\s.