@AlexWaygood 2023-03-10
The `issue-number` check is failing because `gh-102573` does not yet exist as a GitHub issue :)

You should create an issue to describe the change you're proposing here, and then link to it in this PR title.