@XuehaiPan 2023-03-10
Opened a linked issue and added some benchmark results.